requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.